chore: refactor type checking
