+++ packages/monaco/src/editor-hooks.ts
@@ -8,11 +8,11 @@ class EditorEvents extends EventTarget {
 
 const event = new EditorEvents();
 type EventNames = 'resetCode';
-export function useResetEditor(deps: unknown[] = []): {
-  subscribe: (eventName: EventNames, cb: () => void) => void;
+export function useResetEditor(): {
+  subscribe: (eventName: EventNames, cb: () => void, deps: unknown[]) => void;
   dispatch: (eventName: EventNames) => void;
 } {
-  function subscribe(eventName: EventNames, cb: () => void) {
+  function subscribe(eventName: EventNames, cb: () => void, deps: unknown[] = []) {
     useEffect(() => {
       const fnc = () => {
         cb();
