+++ packages/monaco/src/utils.ts
@@ -1,3 +1,5 @@
+import type * as MonacoEditor from 'monaco-editor';
+
 export function getEventDeltas(e: MouseEvent | TouchEvent, origin: { x: number; y: number }) {
   if (e instanceof MouseEvent) {
     return {
@@ -24,4 +26,35 @@ export function getEventDeltas(e: MouseEvent | TouchEvent, origin: { x: number; 
     currPosX: touch.clientX,
     currPosY: touch.clientY,
   };
+}
+
+export async function typeCheck(monaco: typeof MonacoEditor) {
+  const models = monaco.editor.getModels();
+  const getWorker = await monaco.languages.typescript.getTypeScriptWorker();
+
+  for (const model of models) {
+    const worker = await getWorker(model.uri);
+    const diagnostics = (
+      await Promise.all([
+        worker.getSyntacticDiagnostics(model.uri.toString()),
+        worker.getSemanticDiagnostics(model.uri.toString()),
+      ])
+    ).reduce((a, b) => a.concat(b));
+
+    const markers = diagnostics.map((d) => {
+      const start = model.getPositionAt(d.start!);
+      const end = model.getPositionAt(d.start! + d.length!);
+
+      return {
+        severity: monaco.MarkerSeverity.Error,
+        startLineNumber: start.lineNumber,
+        endLineNumber: end.lineNumber,
+        startColumn: start.column,
+        endColumn: end.column,
+        message: d.messageText as string,
+      } satisfies MonacoEditor.editor.IMarkerData;
+    });
+
+    monaco.editor.setModelMarkers(model, model.getLanguageId(), markers);
+  }
 }
